requirements.txt
# local package
-e .

# external requirements
click~=8.0.1
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

agraph-python~=101.0.7
python-docx~=0.8.11
openpyxl~=3.0.7
pandas~=1.3.1
tqdm~=4.61.2
torch~=1.9.0+cu111
numpy~=1.19.5
tensorflow~=2.6.0
transformers~=4.9.0
natasha~=1.4.0
sklearn~=0.0
scikit-learn~=0.24.2
requests~=2.26.0
aiofiles~=0.7.0
Hypercorn~=0.11.2
fastapi~=0.67.0
pydantic~=1.8.2
setuptools~=57.4.0
datefinder~=0.7.1